tweaking fractal land a bit (#491)
* tweaking fractal land a bit

* more aggressive slope

* Use Builder patten for FractalParams and FractalFlattenParams

* Make all the parameters for FractalParams and FractalFlattenParams required with the @Builder pattern

* line indentation and formatting for readability

* remove visual debugger

* Remove @NonNull from FractalParams and FractalFlattenParams, it is not needed.

* Fractal Plateau adjustments

* And a final tweak to fractal plateau.

+++ generator/src/main/java/com/faforever/neroxis/generator/terrain/FractalNoiseLastTerrainGenerator.java
@@ -172,7 +172,7 @@ public class FractalNoiseLastTerrainGenerator extends MultiLevelLastTerrainGener
             if (fractalFlattenParams.spawnable()) {
                 spawnMask.add(landNoiseMap.copyAsBooleanMask(fractalFlattenParams.minHeight(),
                                                              fractalFlattenParams.maxHeight())
-                                          .deflate((int) fractalFlattenParams.spawnMaskDeflate()));
+                                          .deflate(fractalFlattenParams.spawnMaskDeflate()));
             }
         }
